+++ packages/types/src/interfaces/runtime/types.ts
@@ -339,6 +339,15 @@ export interface SlotDuration extends u64 {}
 /** @name StorageData */
 export interface StorageData extends Bytes {}
 
+/** @name StorageInfo */
+export interface StorageInfo extends Struct {
+  readonly palletName: Bytes;
+  readonly storage_name: Bytes;
+  readonly prefix: Bytes;
+  readonly maxValues: Option<u32>;
+  readonly maxSize: Option<u32>;
+}
+
 /** @name StorageProof */
 export interface StorageProof extends Struct {
   readonly trieNodes: Vec<Bytes>;
@@ -352,6 +361,9 @@ export interface TransactionInfo extends Struct {
   readonly blockChunks: u32;
 }
 
+/** @name TransactionLongevity */
+export interface TransactionLongevity extends u64 {}
+
 /** @name TransactionPriority */
 export interface TransactionPriority extends u64 {}
 
@@ -360,6 +372,9 @@ export interface TransactionStorageProof extends Struct {
   readonly chunk: Bytes;
   readonly proof: Vec<Bytes>;
 }
+
+/** @name TransactionTag */
+export interface TransactionTag extends Bytes {}
 
 /** @name U32F32 */
 export interface U32F32 extends UInt {}